Merge from upstream

+++ resource-managers/kubernetes/core/src/main/scala/org/apache/spark/scheduler/cluster/k8s/ExecutorPodsAllocator.scala
@@ -23,7 +23,7 @@ import io.fabric8.kubernetes.api.model.PodBuilder
 import io.fabric8.kubernetes.client.KubernetesClient
 import scala.collection.mutable
 
-import org.apache.spark.{SparkConf, SparkException}
+import org.apache.spark.{SecurityManager, SparkConf, SparkException}
 import org.apache.spark.deploy.k8s.Config._
 import org.apache.spark.deploy.k8s.Constants._
 import org.apache.spark.deploy.k8s.KubernetesConf
@@ -32,6 +32,7 @@ import org.apache.spark.util.{Clock, Utils}
 
 private[spark] class ExecutorPodsAllocator(
     conf: SparkConf,
+    secMgr: SecurityManager,
     executorBuilder: KubernetesExecutorBuilder,
     kubernetesClient: KubernetesClient,
     snapshotsStore: ExecutorPodsSnapshotsStore,
@@ -52,6 +53,8 @@ private[spark] class ExecutorPodsAllocator(
   private val kubernetesDriverPodName = conf
     .get(KUBERNETES_DRIVER_POD_NAME)
 
+  private val shouldDeleteExecutors = conf.get(KUBERNETES_DELETE_EXECUTORS)
+
   private val driverPod = kubernetesDriverPodName
     .map(name => Option(kubernetesClient.pods()
       .withName(name)
@@ -64,10 +67,6 @@ private[spark] class ExecutorPodsAllocator(
   // snapshot yet. Mapped to the timestamp when they were created.
   private val newlyCreatedExecutors = mutable.Map.empty[Long, Long]
 
-  private var latestSnapshot: Option[ExecutorPodsSnapshot] = None
-
-  private var appId: Option[String] = None
-
   def start(applicationId: String): Unit = {
     snapshotsStore.addSubscriber(podAllocationDelay) {
       onNewSnapshots(applicationId, _)
@@ -77,11 +76,9 @@ private[spark] class ExecutorPodsAllocator(
   def setTotalExpectedExecutors(total: Int): Unit = {
     safeLogDebug("Setting total expected executors", SafeArg.of("totalExpectedExecutors", total))
     totalExpectedExecutors.set(total)
-    appId.foreach { id => latestSnapshot.foreach { requestExecutorsIfNecessary(id, _) } }
   }
 
   private def onNewSnapshots(applicationId: String, snapshots: Seq[ExecutorPodsSnapshot]): Unit = {
-    this.appId = Some(applicationId)
     newlyCreatedExecutors --= snapshots.flatMap(_.executorPods.keys)
     // For all executors we've created against the API but have not seen in a snapshot
     // yet - check the current time. If the current time has exceeded some threshold,
@@ -98,11 +95,16 @@ private[spark] class ExecutorPodsAllocator(
           " deleted but the application missed the deletion event.",
           SafeArg.of("executorId", execId),
           SafeArg.of("podCreationTimeoutMs", podCreationTimeout))
-        Utils.tryLogNonFatalError {
-          kubernetesClient
-            .pods()
-            .withLabel(SPARK_EXECUTOR_ID_LABEL, execId.toString)
-            .delete()
+
+        if (shouldDeleteExecutors) {
+          Utils.tryLogNonFatalError {
+            kubernetesClient
+              .pods()
+              .withLabel(SPARK_APP_ID_LABEL, applicationId)
+              .withLabel(SPARK_ROLE_LABEL, SPARK_POD_EXECUTOR_ROLE)
+              .withLabel(SPARK_EXECUTOR_ID_LABEL, execId.toString)
+              .delete()
+          }
         }
         newlyCreatedExecutors -= execId
       } else {
@@ -116,66 +118,60 @@ private[spark] class ExecutorPodsAllocator(
     if (snapshots.nonEmpty) {
       // Only need to examine the cluster as of the latest snapshot, the "current" state, to see if
       // we need to allocate more executors or not.
-      latestSnapshot = Some(snapshots.last)
-      requestExecutorsIfNecessary(applicationId, snapshots.last)
-    }
-  }
-
-  private def requestExecutorsIfNecessary(applicationId: String,
-                                          snapshot: ExecutorPodsSnapshot): Unit = {
-    val currentRunningExecutors = snapshot.executorPods.values.count {
-      case PodRunning(_) => true
-      case _ => false
-    }
-    val currentPendingExecutors = snapshot.executorPods.values.count {
-      case PodPending(_) => true
-      case _ => false
-    }
-    val currentTotalExpectedExecutors = totalExpectedExecutors.get
-    safeLogDebug("Currently have running executors and" +
-      " pending executors. Newly created executors" +
-      " have been requested but are pending appearance in the cluster.",
-      SafeArg.of("numCurrentRunningExecutors", currentRunningExecutors),
-      SafeArg.of("numCurrentPendingExecutors", currentPendingExecutors),
-      SafeArg.of("newlyCreatedExecutors", newlyCreatedExecutors))
-    if (newlyCreatedExecutors.isEmpty
-      && currentPendingExecutors == 0
-      && currentRunningExecutors < currentTotalExpectedExecutors) {
-      val numExecutorsToAllocate = math.min(
-        currentTotalExpectedExecutors - currentRunningExecutors, podAllocationSize)
-      safeLogInfo("Going to request executors from Kubernetes.",
-        SafeArg.of("numExecutorsToAllocate", numExecutorsToAllocate))
-      for ( _ <- 0 until numExecutorsToAllocate) {
-        val newExecutorId = EXECUTOR_ID_COUNTER.incrementAndGet()
-        val executorConf = KubernetesConf.createExecutorConf(
-          conf,
-          newExecutorId.toString,
-          applicationId,
-          driverPod)
-        val executorPod = executorBuilder.buildFromFeatures(executorConf)
-        val podWithAttachedContainer = new PodBuilder(executorPod.pod)
-          .editOrNewSpec()
-          .addToContainers(executorPod.container)
-          .endSpec()
-          .build()
-        kubernetesClient.pods().create(podWithAttachedContainer)
-        newlyCreatedExecutors(newExecutorId) = clock.getTimeMillis()
-        safeLogDebug("Requested executor from Kubernetes.",
-          SafeArg.of("newExecutorId", newExecutorId))
+      val latestSnapshot = snapshots.last
+      val currentRunningExecutors = latestSnapshot.executorPods.values.count {
+        case PodRunning(_) => true
+        case _ => false
       }
-    } else if (currentRunningExecutors >= currentTotalExpectedExecutors) {
-      // TODO handle edge cases if we end up with more running executors than expected.
-      safeLogDebug("Current number of running executors is equal to the number of requested" +
-        " executors. Not scaling up further.")
-    } else if (newlyCreatedExecutors.nonEmpty || currentPendingExecutors != 0) {
-      safeLogDebug("Still waiting for" +
-        " executors to begin running before requesting for more executors, including executors" +
-        " in pending status in the cluster, and executors that we have" +
-        " created but we have not observed as being present in the cluster yet.",
-        SafeArg.of("numTotalCurrentWaitingExecutors",
-          newlyCreatedExecutors.size + currentPendingExecutors),
-        SafeArg.of("numCurrentPendingExecutors", currentPendingExecutors),
-        SafeArg.of("numNewlyCreatedExecutors", newlyCreatedExecutors.size))
+      val currentPendingExecutors = latestSnapshot.executorPods.values.count {
+        case PodPending(_) => true
+        case _ => false
+      }
+      val currentTotalExpectedExecutors = totalExpectedExecutors.get
+        safeLogDebug("Currently have running executors and" +
+          " pending executors. Newly created executors" +
+          " have been requested but are pending appearance in the cluster.",
+          SafeArg.of("numCurrentRunningExecutors", currentRunningExecutors),
+          SafeArg.of("numCurrentPendingExecutors", currentPendingExecutors),
+          SafeArg.of("newlyCreatedExecutors", newlyCreatedExecutors))
+      if (newlyCreatedExecutors.isEmpty
+        && currentPendingExecutors == 0
+        && currentRunningExecutors < currentTotalExpectedExecutors) {
+        val numExecutorsToAllocate = math.min(
+          currentTotalExpectedExecutors - currentRunningExecutors, podAllocationSize)
+        safeLogInfo("Going to request executors from Kubernetes.",
+          SafeArg.of("numExecutorsToAllocate", numExecutorsToAllocate))
+        for ( _ <- 0 until numExecutorsToAllocate) {
+          val newExecutorId = EXECUTOR_ID_COUNTER.incrementAndGet()
+          val executorConf = KubernetesConf.createExecutorConf(
+            conf,
+            newExecutorId.toString,
+            applicationId,
+            driverPod)
+          val executorPod = executorBuilder.buildFromFeatures(executorConf, secMgr,
+            kubernetesClient)
+          val podWithAttachedContainer = new PodBuilder(executorPod.pod)
+            .editOrNewSpec()
+            .addToContainers(executorPod.container)
+            .endSpec()
+            .build()
+          kubernetesClient.pods().create(podWithAttachedContainer)
+          newlyCreatedExecutors(newExecutorId) = clock.getTimeMillis()
+          safeLogDebug("Requested executor from Kubernetes.",
+            SafeArg.of("newExecutorId", newExecutorId))
+        }
+      } else if (currentRunningExecutors >= currentTotalExpectedExecutors) {
+        // TODO handle edge cases if we end up with more running executors than expected.
+        safeLogDebug("Current number of running executors is equal to the number of requested" +
+          " executors. Not scaling up further.")
+      } else if (newlyCreatedExecutors.nonEmpty || currentPendingExecutors != 0) {
+        safeLogDebug("Still waiting for executors to begin running before requesting for more " +
+          " executors.",
+          SafeArg.of("numWaitingForExecutors",
+            newlyCreatedExecutors.size + currentPendingExecutors),
+          SafeArg.of("numPendingExecutors", currentPendingExecutors),
+          SafeArg.of("numNewlyCreatedExecutors", newlyCreatedExecutors.size))
+      }
     }
   }
 }
